I have added doc strings in the login and user registration classes to establish clarity of their functions and roles. I also refactored LoginAction into its own LoginAction package in the UseCase package to follow the architecture of the Use Case level of our project.
