+++ src/main/java/UseCases/UserRegister/UserRegisterCreateUser.java
@@ -1,5 +1,8 @@
 package UseCases.UserRegister;
 
+/**
+ * An interface to define a "create" functionality for classes to create a UserRegisterInputs
+ */
 public interface UserRegisterCreateUser {
     UserRegisterInputs create(UserRegisterInputs inputs);
 }
