+++ src/main/java/UseCases/UserRegister/UserRegisterStatus.java
@@ -1,7 +1,8 @@
 package UseCases.UserRegister;
 
-import UseCases.UserRegister.UserRegisterInputs;
-
+/**
+ * An interface to define the functionality of successes and failures of creating a new user
+ */
 public interface UserRegisterStatus {
     UserRegisterInputs showSuccess(UserRegisterInputs newUser);
     UserRegisterInputs showFailure(String failMessage);
